safari-600.7-branch/LayoutTests/platform/ios-sim/fast/forms/search-disabled-readonly-expected.txt




This tests the behavior of a cancel button in search input forms.


Test on the input form with disabled=false and readonly=false
Click the cancel button:
FAIL input.value should be . Was foo.
... and then input one character:
FAIL input.value should be b. Was foo.
Click the center of the form:
PASS input.value is "foo"
... and then input one character:
FAIL input.value should be foob. Was foo.

Test on the input form with disabled=false and readonly=true
Click the cancel button:
PASS input.value is "foo"
... and then input one character:
PASS input.value is "foo"
Click the center of the form:
PASS input.value is "foo"
... and then input one character:
PASS input.value is "foo"

Test on the input form with disabled=true and readonly=false
Click the cancel button:
PASS input.value is "foo"
... and then input one character:
PASS input.value is "foo"
Click the center of the form:
PASS input.value is "foo"
... and then input one character:
PASS input.value is "foo"

Test on the input form with disabled=true and readonly=true
Click the cancel button:
PASS input.value is "foo"
... and then input one character:
PASS input.value is "foo"
Click the center of the form:
PASS input.value is "foo"
... and then input one character:
PASS input.value is "foo"

PASS successfullyParsed is true

TEST COMPLETE
